+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Update rspec to `2.99.0.beta1` on Gemfile
 - Use FactoryGirl.lint to test all factories before running specs
 - Update rails version to ```4.0.3```
+- Run `rails generate simple_form:install`
 
 ### bug fixes